extract input components to its own folder in components
